requirements.txt
behave
pytest-playwright
allure-behave